rename samples to samplelib
