+++ samplelib/SampleGeneratorBase.py
